hiding things away from package object
