--- src/main/scala/wdl4s/cwl/RequirementClass.scala
@@ -1,7 +1,0 @@
-package wdl4s.cwl
-
-object RequirementClass extends Enumeration {
-  type RequirementClass = Value
-
-  val EnvVarRequirementClass = Value("EnvVarRequirement")
-}
