+++ src/main/scala/wdl4s/cwl/model.scala
@@ -4,7 +4,6 @@ import eu.timepit.refined._
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.string._
 import shapeless.{:+:, CNil, Witness}
-import RequirementClass._
 
 case class WorkflowStepInput(src: String)
